@@ -2,13 +2,15 @@
   "name": "Airbnb-booking",
   "version": "1.0.0",
   "description": "> Airbnb Booking module",
-  "main": "index.js",
+  "main": "server/index.js",
   "scripts": {
     "test": "jest --coverage",
     "start": "nodemon server/index.js",
     "build": "webpack --config webpack.config.prod.js",
     "react-dev": "webpack -d --watch",
-    "seed": "mysql -u root <db/schema.sql && node db/room_generater.js && db/booking_generater.js"
+    "react-prod": "webpack -p",
+    "seed": "mysql -u root <db/schema.sql && node db/room_generater.js && db/booking_generater.js",
+    "ssr": "webpack --config webpack.config.prod.js & nodemon server/server.js"
   },
   "lint-staged": {
     "*.js": [
@@ -34,48 +36,71 @@
   },
   "homepage": "https://github.com/hacker-home/Airbnb-booking#readme",
   "dependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "@newrelic/native-metrics": "^5.3.0",
+    "axios": "^0.20.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^23.4.2",
     "body-parser": "^1.19.0",
+    "compression": "^1.7.4",
     "cors": "^2.8.5",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.13.1",
-    "express": "^4.17.0",
+    "dotenv": "^8.2.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.4",
+    "express": "^4.17.1",
+    "express-static-gzip": "^2.0.8",
     "faker": "^4.1.0",
-    "jest": "^24.8.0",
-    "jquery": "^3.5.0",
-    "moment": "^2.24.0",
-    "mysql": "^2.17.1",
-    "mysql2": "^1.6.5",
-    "nodemon": "^1.19.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "sequelize": "^5.15.1",
-    "prop-types": "^15.7.2"
+    "ignore-styles": "^5.0.1",
+    "iltorb": "^2.4.5",
+    "jest": "^24.9.0",
+    "moment": "^2.27.0",
+    "morgan": "^1.10.0",
+    "mysql": "^2.18.1",
+    "mysql2": "^1.7.0",
+    "newrelic": "^6.13.0",
+    "nodemon": "^1.19.4",
+    "pg": "^8.3.3",
+    "pg-hstore": "^2.3.3",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "redis": "^3.0.2",
+    "sequelize": "^5.22.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.4",
-    "@babel/preset-env": "^7.4.4",
-    "@babel/preset-react": "^7.0.0",
-    "babel": "^6.23.0",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.11.6",
+    "@babel/plugin-proposal-export-default-from": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^23.4.2",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-css-modules-transform": "^1.6.2",
+    "brotli-webpack-plugin": "^1.1.0",
     "compression-webpack-plugin": "^2.0.0",
     "css-loader": "^2.1.1",
     "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.13.0",
-    "husky": "^2.3.0",
+    "eslint-config-airbnb": "^17.1.1",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react": "^7.20.6",
+    "husky": "^2.7.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest-cli": "^24.8.0",
-    "lint-staged": "^8.1.7",
+    "jest-cli": "^24.9.0",
+    "lint-staged": "^8.2.1",
     "pre-commit": "^1.2.2",
-    "react-test-renderer": "^16.8.6",
-    "sinon": "^7.3.2",
+    "react-test-renderer": "^16.13.1",
+    "sinon": "^7.5.0",
     "style-loader": "^0.13.2",
     "typescript-babel-jest": "^1.0.6",
-    "webpack": "^4.31.0",
-    "webpack-cli": "^3.3.2",
-    "webpack-merge": "^4.2.1"
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12",
+    "webpack-merge": "^4.2.2",
+    "webpack-node-externals": "^2.5.2"
   }
 }
