@@ -1,10 +1,11 @@
 /* eslint-disable import/extensions */
 import React from 'react';
-import $ from 'jquery';
 import moment from 'moment';
 import Info from './components/Info.jsx';
 import Form from './components/Form.jsx';
 import css from '../../public/dist/App.css';
+import axios from 'axios';
+import idGenerator from '../../biasedNumGenerator.js';
 
 export default class App extends React.Component {
   constructor(props) {
@@ -23,6 +24,7 @@ export default class App extends React.Component {
         ratings: '',
         numReviews: 0,
       },
+      biasedNum: idGenerator(),
       bookedDates: [],
       rendering: true,
     };
@@ -42,54 +44,34 @@ export default class App extends React.Component {
   getRoomData() {
     const link = window.location.href.match(/id\s*=\s*(.*)/);
     if (link) {
-      $.ajax({
-        url: `/room?id=${link[1]}`,
-        type: 'GET',
-        error: (err) => {
-          console.log(err);
-        },
-        success: (result) => {
-          this.updateRoomState(result);
-        },
-      });
+      axios.get(`/room?id=${link[1]}`)
+        .then(({ data }) => {
+          this.updateRoomState(data);
+        })
+        .catch(err => console.log(err))
     } else {
-      $.ajax({
-        url: '/room?id=1',
-        type: 'GET',
-        error: (err) => {
-          console.log(err);
-        },
-        success: (result) => {
-          this.updateRoomState(result);
-        },
-      });
+      axios.get(`/room?id=${this.state.biasedNum}`)
+        .then(({ data }) => {
+          this.updateRoomState(data);
+        })
+        .catch(err => console.log(err))
     }
   }
 
   getBookingData() {
     const link = window.location.href.match(/id\s*=\s*(.*)/);
     if (link) {
-      $.ajax({
-        url: `/booking?id=${link[1]}`,
-        type: 'GET',
-        error: (err) => {
-          console.log(err);
-        },
-        success: (result) => {
-          this.updateBookedDates(result);
-        },
-      });
+      axios.get(`/booking?id=${link[1]}`)
+        .then(({ data }) => {
+          this.updateBookedDates(data);
+        })
+        .catch(err => console.log(err))
     } else {
-      $.ajax({
-        url: '/booking?id=1',
-        type: 'GET',
-        error: (err) => {
-          console.log(err);
-        },
-        success: (result) => {
-          this.updateBookedDates(result);
-        },
-      });
+      axios.get(`/booking?id=${this.state.biasedNum}`)
+        .then(({ data }) => {
+          this.updateBookedDates(data);
+        })
+        .catch(err => console.log(err))
     }
   }
 
@@ -118,14 +100,20 @@ export default class App extends React.Component {
   }
 
   updateRoomState(result) {
+    const maxGuest = {
+      adults: result.max_adults,
+      children: result.max_children,
+      infants: result.max_infants
+    }
     this.setState({
+      roomId: result.id,
       roomInfo: {
         roomname: result.roomname,
         price: result.price,
         cleaningFee: result.cleaning_fee,
         serviceFee: result.service_fee,
         tax: result.tax,
-        maxGuest: result.max_guest,
+        maxGuest: maxGuest,
         minNight: result.min_night,
         maxNight: result.max_night,
         ratings: result.ratings,
