@@ -3,4 +3,5 @@ import ReactDOM from 'react-dom';
 import App from './App.jsx';
 
 
-ReactDOM.render(<App />, document.getElementById('booking'));
+ReactDOM.hydrate(<App />, document.getElementById('booking'));
+
